+++ lib/service/run.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/Jeffail/benthos/v3/lib/config"
 	"github.com/Jeffail/benthos/v3/lib/input"
-	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/output"
 	"github.com/Jeffail/benthos/v3/lib/processor"
+	"github.com/Jeffail/benthos/v3/lib/service/blobl"
 	"github.com/Jeffail/benthos/v3/lib/service/test"
 	uconfig "github.com/Jeffail/benthos/v3/lib/util/config"
 	"github.com/urfave/cli/v2"
@@ -137,6 +137,12 @@ func Run() {
 	app := &cli.App{
 		Name:  "benthos",
 		Usage: "A stream processor for mundane tasks - https://benthos.dev",
+		Description: `
+   Either run Benthos as a stream processor or choose a command:
+
+   benthos list inputs
+   benthos create kafka_balanced//file > ./config.yaml
+   benthos -c ./config.yaml`[4:],
 		Flags: []cli.Flag{
 			&cli.BoolFlag{
 				Name:    "version",
@@ -313,57 +319,8 @@ func Run() {
 					return nil
 				},
 			},
-			{
-				Name:  "test",
-				Usage: "Execute Benthos unit tests",
-				Description: `
-   Execute any number of Benthos unit test definitions. If one or more tests
-   fail the process will report the errors and exit with a status code 1.
-
-   benthos test ./path/to/configs/...
-   benthos test ./foo_configs ./bar_configs
-   benthos test ./foo.yaml
-
-   For more information check out the docs at:
-   https://benthos.dev/docs/configuration/unit_testing`[4:],
-				Flags: []cli.Flag{
-					&cli.BoolFlag{
-						Name:  "generate",
-						Value: false,
-						Usage: "instead of testing, detect untested Benthos configs and generate test definitions for them.",
-					},
-					&cli.StringFlag{
-						Name:  "log",
-						Value: "",
-						Usage: "allow components to write logs at a provided level to stdout.",
-					},
-				},
-				Action: func(c *cli.Context) error {
-					if c.Bool("generate") {
-						if err := test.GenerateAll(
-							c.Args().Slice(), testSuffix,
-						); err != nil {
-							fmt.Fprintf(os.Stderr, "Failed to generate config tests: %v\n", err)
-							os.Exit(1)
-						}
-						os.Exit(0)
-					}
-					if logLevel := c.String("log"); len(logLevel) > 0 {
-						logConf := log.NewConfig()
-						logConf.LogLevel = logLevel
-						logger := log.New(os.Stdout, logConf)
-						if test.RunAllWithLogger(c.Args().Slice(), testSuffix, true, logger) {
-							os.Exit(0)
-						}
-					} else {
-						if test.RunAll(c.Args().Slice(), testSuffix, true) {
-							os.Exit(0)
-						}
-					}
-					os.Exit(1)
-					return nil
-				},
-			},
+			test.CliCommand(testSuffix),
+			blobl.CliCommand(),
 		},
 	}